add tests for state.py (partial)
- added some basic tests for state.py in test_state.py
- infrastructure to test for product upgrades (only for state.py for now):
  * added create_migration_db.py script and containers.json, pkgs.json to be able to create
    the state database for any released version so far
  * generated state database for 0.9.0, 0.9.1, 0.9.2, 0.9.5, 0.9.6 product versions that
    have seen schema changes
- added test_migration in test_state.py to check for migration of all the above mentioned
  versions, and confirming proper data for the same
- most other tests are just basic shell which will be updated in upcoming checkins
- explicitly ensure mode=0750 for all directories created by implicitly to ensure better
  permissions

+++ src/ybox/state.py
@@ -4,6 +4,7 @@ installed on each container explicitly.
 """
 
 import json
+import os
 import re
 import sqlite3
 from configparser import ConfigParser
@@ -21,7 +22,7 @@ from packaging.version import parse as parse_version
 
 from ybox import __version__ as product_version
 
-from .config import StaticConfiguration
+from .config import Consts, StaticConfiguration
 from .env import Environ, PathName
 from .print import print_color, print_warn
 from .util import ini_file_reader, resolve_inc_path, write_ybox_version
@@ -126,6 +127,7 @@ class YboxStateManagement:
         """
         # explicitly control transaction begin (in exclusive mode) since SERIALIZABLE isolation
         # level is required while sqlite3 module will not start transactions before reads
+        os.makedirs(env.data_dir, mode=Consts.default_directory_mode(), exist_ok=True)
         self._conn = sqlite3.connect(f"{env.data_dir}/state.db", timeout=60,
                                      isolation_level=None)
         self._explicit_transaction = False
@@ -354,8 +356,8 @@ class YboxStateManagement:
         write_ybox_version(conf)
 
     def register_container(self, container_name: str, distribution: str, shared_root: str,
-                           parser: ConfigParser,
-                           force_own_orphans: bool) -> dict[str, tuple[CopyType, dict[str, str]]]:
+                           parser: ConfigParser, force_own_orphans: bool = True) -> \
+            dict[str, tuple[CopyType, dict[str, str]]]:
         """
         Register information of a ybox container including its name, distribution and
         configuration. In addition to the registration, this will also check for orphaned packages
@@ -548,7 +550,10 @@ class YboxStateManagement:
         """
         if shared_root:
             shared_containers = self.get_containers(shared_root=shared_root)
-            shared_containers.remove(container_name)
+            try:
+                shared_containers.remove(container_name)
+            except ValueError:
+                pass
             return shared_containers
         return []
 
@@ -793,8 +798,9 @@ class YboxStateManagement:
     def _remove_local_copies(local_copies: list[str]) -> None:
         """remove the files created locally to run container executables"""
         for file in local_copies:
-            print_warn(f"Removing local wrapper/link {file}")
-            Path(file).unlink(missing_ok=True)
+            if os.path.exists(file):
+                print_warn(f"Removing local wrapper/link {file}")
+                Path(file).unlink(missing_ok=True)
 
     def get_packages(self, container_name: str, regex: str = ".*",
                      dependency_type: str = ".*") -> list[str]:
